requirements.txt
Django==3.2
djangorestframework==3.12.4
psycopg2==2.8.6
django-cors-headers==3.7.0
pandas==1.2.4
numpy==1.20.2
coreapi>=2.3.3
pyyaml==5.4.1
drf-yasg[validation]==1.20.0
